+++ packages/store/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk-tools/store
 
+## 1.0.0
+
+### Minor Changes
+
+- Synchronize all package versions.
+
+## 0.9.1
+
+### Patch Changes
+
+- Routine release.
+
 ## 0.7.0
 
 ### Minor Changes